mymodel/requirements.txt
fastapi
uvicorn[standard]
# tensorflow
tensorflow_hub
pydantic
httpx
pytest